@@ -1,6 +1,6 @@
 {
   "name": "jsoneditor",
-  "version": "5.1.5",
+  "version": "8.1.0",
   "main": "./index",
   "description": "A web-based tool to view, edit, format, and validate JSON",
   "tags": [
@@ -19,23 +19,57 @@
   "bugs": "https://github.com/josdejong/jsoneditor/issues",
   "scripts": {
     "build": "gulp",
-    "watch": "watch",
-    "test": "mocha test"
+    "minify": "gulp minify",
+    "start": "gulp watch",
+    "test": "mocha test --require @babel/register",
+    "lint": "standard --env=mocha",
+    "prepublishOnly": "npm test && npm run build"
   },
   "dependencies": {
-    "ajv": "3.4.0",
-    "brace": "0.7.0"
+    "ace-builds": "^1.4.7",
+    "ajv": "^6.10.2",
+    "javascript-natural-sort": "^0.7.1",
+    "jmespath": "^0.15.0",
+    "json-source-map": "^0.6.1",
+    "mobius1-selectr": "^2.4.13",
+    "picomodal": "^3.0.0",
+    "vanilla-picker": "^2.10.0"
   },
   "devDependencies": {
-    "gulp": "3.9.0",
-    "gulp-concat-css": "2.2.0",
-    "gulp-minify-css": "1.2.3",
-    "gulp-shell": "0.5.1",
-    "gulp-util": "3.0.7",
-    "json-loader": "0.5.4",
-    "mkdirp": "0.5.1",
-    "mocha": "2.3.4",
-    "uglify-js": "2.6.1",
-    "webpack": "1.12.11"
+    "@babel/core": "^7.7.5",
+    "@babel/preset-env": "^7.7.6",
+    "@babel/register": "^7.7.4",
+    "babel-loader": "^8.0.6",
+    "btoa": "^1.2.1",
+    "date-format": "^3.0.0",
+    "fancy-log": "^1.3.3",
+    "gulp": "^4.0.2",
+    "gulp-clean-css": "^4.2.0",
+    "gulp-concat-css": "^3.1.0",
+    "gulp-sass": "^4.0.2",
+    "jsdom": "^15.2.1",
+    "json-loader": "^0.5.7",
+    "mkdirp": "^0.5.1",
+    "mocha": "^6.2.2",
+    "standard": "^14.3.1",
+    "uglify-js": "^3.7.2",
+    "webpack": "^4.41.2"
+  },
+  "files": [
+    "dist",
+    "docs",
+    "examples",
+    "src",
+    "HISTORY.md",
+    "index.js",
+    "LICENSE",
+    "NOTICE",
+    "README.md"
+  ],
+  "standard": {
+    "ignore": [
+      "src/js/assets",
+      "examples/react*"
+    ]
   }
 }
