@@ -1,4 +1,13 @@
-var jsonlint = require('./assets/jsonlint/jsonlint');
+'use strict'
+
+import './polyfills'
+import naturalSort from 'javascript-natural-sort'
+import jsonlint from './assets/jsonlint/jsonlint'
+import jsonMap from 'json-source-map'
+import { translate } from './i18n'
+
+const MAX_ITEMS_FIELDS_COLLECTION = 10000
+const YEAR_2000 = 946684800000
 
 /**
  * Parse JSON using the parser built-in in the browser.
@@ -6,154 +15,263 @@ var jsonlint = require('./assets/jsonlint/jsonlint');
  * @param {String} jsonString
  * @return {JSON} json
  */
-exports.parse = function parse(jsonString) {
+export function parse (jsonString) {
   try {
-    return JSON.parse(jsonString);
-  }
-  catch (err) {
+    return JSON.parse(jsonString)
+  } catch (err) {
     // try to throw a more detailed error message using validate
-    exports.validate(jsonString);
+    validate(jsonString)
 
     // rethrow the original error
-    throw err;
+    throw err
   }
-};
+}
 
 /**
- * Sanitize a JSON-like string containing. For example changes JavaScript
+ * Repair a JSON-like string containing. For example changes JavaScript
  * notation into JSON notation.
  * This function for example changes a string like "{a: 2, 'b': {c: 'd'}"
  * into '{"a": 2, "b": {"c": "d"}'
  * @param {string} jsString
  * @returns {string} json
  */
-exports.sanitize = function (jsString) {
-  // escape all single and double quotes inside strings
-  var chars = [];
-  var i = 0;
+export function repair (jsString) {
+  // TODO: refactor this function, it's too large and complicated now
 
-  //If JSON starts with a function (characters/digits/"_-"), remove this function.
-  //This is useful for "stripping" JSONP objects to become JSON
-  //For example: /* some comment */ function_12321321 ( [{"a":"b"}] ); => [{"a":"b"}]
-  var match = jsString.match(/^\s*(\/\*(.|[\r\n])*?\*\/)?\s*[\da-zA-Z_$]+\s*\(([\s\S]*)\)\s*;?\s*$/);
+  // escape all single and double quotes inside strings
+  const chars = []
+  let i = 0
+
+  // If JSON starts with a function (characters/digits/"_-"), remove this function.
+  // This is useful for "stripping" JSONP objects to become JSON
+  // For example: /* some comment */ function_12321321 ( [{"a":"b"}] ); => [{"a":"b"}]
+  const match = jsString.match(/^\s*(\/\*(.|[\r\n])*?\*\/)?\s*[\da-zA-Z_$]+\s*\(([\s\S]*)\)\s*;?\s*$/)
   if (match) {
-    jsString = match[3];
+    jsString = match[3]
   }
 
+  const controlChars = {
+    '\b': '\\b',
+    '\f': '\\f',
+    '\n': '\\n',
+    '\r': '\\r',
+    '\t': '\\t'
+  }
+
+  const quote = '\''
+  const quoteDbl = '"'
+  const quoteLeft = '\u2018'
+  const quoteRight = '\u2019'
+  const quoteDblLeft = '\u201C'
+  const quoteDblRight = '\u201D'
+  const graveAccent = '\u0060'
+  const acuteAccent = '\u00B4'
+
   // helper functions to get the current/prev/next character
-  function curr () { return jsString.charAt(i);     }
-  function next()  { return jsString.charAt(i + 1); }
-  function prev()  { return jsString.charAt(i - 1); }
+  function curr () { return jsString.charAt(i) }
+  function next () { return jsString.charAt(i + 1) }
+  function prev () { return jsString.charAt(i - 1) }
+
+  function isWhiteSpace (c) {
+    return c === ' ' || c === '\n' || c === '\r' || c === '\t'
+  }
 
   // get the last parsed non-whitespace character
   function lastNonWhitespace () {
-    var p = chars.length - 1;
+    let p = chars.length - 1
 
     while (p >= 0) {
-      var pp = chars[p];
-      if (pp !== ' ' && pp !== '\n' && pp !== '\r' && pp !== '\t') { // non whitespace
-        return pp;
+      const pp = chars[p]
+      if (!isWhiteSpace(pp)) {
+        return pp
       }
-      p--;
+      p--
     }
 
-    return '';
+    return ''
+  }
+
+  // get at the first next non-white space character
+  function nextNonWhiteSpace () {
+    let iNext = i + 1
+    while (iNext < jsString.length && isWhiteSpace(jsString[iNext])) {
+      iNext++
+    }
+
+    return jsString[iNext]
   }
 
   // skip a block comment '/* ... */'
   function skipBlockComment () {
-    i += 2;
+    i += 2
     while (i < jsString.length && (curr() !== '*' || next() !== '/')) {
-      i++;
+      i++
     }
-    i += 2;
+    i += 2
   }
 
   // skip a comment '// ...'
   function skipComment () {
-    i += 2;
+    i += 2
     while (i < jsString.length && (curr() !== '\n')) {
-      i++;
+      i++
     }
   }
 
-  // parse single or double quoted string
-  function parseString(quote) {
-    chars.push('"');
-    i++;
-    var c = curr();
-    while (i < jsString.length && c !== quote) {
+  /**
+   * parse single or double quoted string. Returns the parsed string
+   * @param {string} endQuote
+   * @return {string}
+   */
+  function parseString (endQuote) {
+    let string = ''
+
+    string += '"'
+    i++
+    let c = curr()
+    while (i < jsString.length && c !== endQuote) {
       if (c === '"' && prev() !== '\\') {
         // unescaped double quote, escape it
-        chars.push('\\');
-      }
-
-      // handle escape character
-      if (c === '\\') {
-        i++;
-        c = curr();
-
+        string += '\\"'
+      } else if (c in controlChars) {
+        // replace unescaped control characters with escaped ones
+        string += controlChars[c]
+      } else if (c === '\\') {
         // remove the escape character when followed by a single quote ', not needed
+        i++
+        c = curr()
         if (c !== '\'') {
-          chars.push('\\');
+          string += '\\'
         }
+        string += c
+      } else {
+        // regular character
+        string += c
       }
-      chars.push(c);
 
-      i++;
-      c = curr();
+      i++
+      c = curr()
     }
-    if (c === quote) {
-      chars.push('"');
-      i++;
+    if (c === endQuote) {
+      string += '"'
+      i++
     }
+
+    return string
   }
 
   // parse an unquoted key
-  function parseKey() {
-    var specialValues = ['null', 'true', 'false'];
-    var key = '';
-    var c = curr();
+  function parseKey () {
+    const specialValues = ['null', 'true', 'false']
+    let key = ''
+    let c = curr()
 
-    var regexp = /[a-zA-Z_$\d]/; // letter, number, underscore, dollar character
+    const regexp = /[a-zA-Z_$\d]/ // letter, number, underscore, dollar character
     while (regexp.test(c)) {
-      key += c;
-      i++;
-      c = curr();
+      key += c
+      i++
+      c = curr()
     }
 
     if (specialValues.indexOf(key) === -1) {
-      chars.push('"' + key + '"');
-    }
-    else {
-      chars.push(key);
+      return '"' + key + '"'
+    } else {
+      return key
     }
   }
 
-  while(i < jsString.length) {
-    var c = curr();
+  function parseMongoDataType () {
+    let c = curr()
+    let value
+    let dataType = ''
+    while (/[a-zA-Z_$]/.test(c)) {
+      dataType += c
+      i++
+      c = curr()
+    }
+
+    if (dataType.length > 0 && c === '(') {
+      // This is an MongoDB data type like {"_id": ObjectId("123")}
+      i++
+      c = curr()
+      if (c === '"') {
+        // a data type containing a string, like ISODate("2012-12-19T06:01:17.171Z")
+        value = parseString(c)
+        c = curr()
+      } else {
+        // a data type containing a value, like 'NumberLong(2)'
+        value = ''
+        while (c !== ')' && c !== '') {
+          value += c
+          i++
+          c = curr()
+        }
+      }
+
+      if (c === ')') {
+        // skip the closing bracket at the end
+        i++
+
+        // return the value (strip the data type object)
+        return value
+      } else {
+        // huh? that's unexpected. don't touch it
+        return dataType + '(' + value + c
+      }
+    } else {
+      // hm, no Mongo data type after all
+      return dataType
+    }
+  }
+
+  function isSpecialWhiteSpace (c) {
+    return (
+      c === '\u00A0' ||
+        (c >= '\u2000' && c <= '\u200A') ||
+        c === '\u202F' ||
+        c === '\u205F' ||
+        c === '\u3000')
+  }
+
+  while (i < jsString.length) {
+    const c = curr()
 
     if (c === '/' && next() === '*') {
-      skipBlockComment();
-    }
-    else if (c === '/' && next() === '/') {
-      skipComment();
-    }
-    else if (c === '\'' || c === '"') {
-      parseString(c);
-    }
-    else if (/[a-zA-Z_$]/.test(c) && ['{', ','].indexOf(lastNonWhitespace()) !== -1) {
+      skipBlockComment()
+    } else if (c === '/' && next() === '/') {
+      skipComment()
+    } else if (isSpecialWhiteSpace(c)) {
+      // special white spaces (like non breaking space)
+      chars.push(' ')
+      i++
+    } else if (c === quote) {
+      chars.push(parseString(c))
+    } else if (c === quoteDbl) {
+      chars.push(parseString(quoteDbl))
+    } else if (c === graveAccent) {
+      chars.push(parseString(acuteAccent))
+    } else if (c === quoteLeft) {
+      chars.push(parseString(quoteRight))
+    } else if (c === quoteDblLeft) {
+      chars.push(parseString(quoteDblRight))
+    } else if (c === ',' && [']', '}'].indexOf(nextNonWhiteSpace()) !== -1) {
+      // skip trailing commas
+      i++
+    } else if (/[a-zA-Z_$]/.test(c) && ['{', ','].indexOf(lastNonWhitespace()) !== -1) {
       // an unquoted object key (like a in '{a:2}')
-      parseKey();
-    }
-    else {
-      chars.push(c);
-      i++;
+      chars.push(parseKey())
+    } else {
+      if (/[a-zA-Z_$]/.test(c)) {
+        chars.push(parseMongoDataType())
+      } else {
+        chars.push(c)
+        i++
+      }
     }
   }
 
-  return chars.join('');
-};
+  return chars.join('')
+}
 
 /**
  * Escape unicode characters.
@@ -161,14 +279,19 @@ exports.sanitize = function (jsString) {
  * @param {string} text
  * @return {string}
  */
-exports.escapeUnicodeChars = function (text) {
+export function escapeUnicodeChars (
   // see https://www.wikiwand.com/en/UTF-16
-  // note: we leave surrogate pairs as two individual chars,
-  // as JSON doesn't interpret them as a single unicode char.
-  return text.replace(/[\u007F-\uFFFF]/g, function(c) {
-    return '\\u'+('0000' + c.charCodeAt(0).toString(16)).slice(-4);
-  })
-};
+  text
+) {
+  return (
+    // note: we leave surrogate pairs as two individual chars,
+    // as JSON doesn't interpret them as a single unicode char.
+    text.replace(
+      /[\u007F-\uFFFF]/g,
+      c => '\\u' + ('0000' + c.charCodeAt(0).toString(16)).slice(-4)
+    )
+  )
+}
 
 /**
  * Validate a string containing a JSON object
@@ -177,14 +300,13 @@ exports.escapeUnicodeChars = function (text) {
  * @param {String} jsonString   String with an (invalid) JSON object
  * @throws Error
  */
-exports.validate = function validate(jsonString) {
-  if (typeof(jsonlint) != 'undefined') {
-    jsonlint.parse(jsonString);
+export function validate (jsonString) {
+  if (typeof (jsonlint) !== 'undefined') {
+    jsonlint.parse(jsonString)
+  } else {
+    JSON.parse(jsonString)
   }
-  else {
-    JSON.parse(jsonString);
-  }
-};
+}
 
 /**
  * Extend object a with the properties of object b
@@ -192,79 +314,80 @@ exports.validate = function validate(jsonString) {
  * @param {Object} b
  * @return {Object} a
  */
-exports.extend = function extend(a, b) {
-  for (var prop in b) {
-    if (b.hasOwnProperty(prop)) {
-      a[prop] = b[prop];
+export function extend (a, b) {
+  for (const prop in b) {
+    if (hasOwnProperty(b, prop)) {
+      a[prop] = b[prop]
     }
   }
-  return a;
-};
+  return a
+}
 
 /**
  * Remove all properties from object a
  * @param {Object} a
  * @return {Object} a
  */
-exports.clear = function clear (a) {
-  for (var prop in a) {
-    if (a.hasOwnProperty(prop)) {
-      delete a[prop];
+export function clear (a) {
+  for (const prop in a) {
+    if (hasOwnProperty(a, prop)) {
+      delete a[prop]
     }
   }
-  return a;
-};
+  return a
+}
 
 /**
  * Get the type of an object
  * @param {*} object
  * @return {String} type
  */
-exports.type = function type (object) {
+export function getType (object) {
   if (object === null) {
-    return 'null';
+    return 'null'
   }
   if (object === undefined) {
-    return 'undefined';
+    return 'undefined'
   }
   if ((object instanceof Number) || (typeof object === 'number')) {
-    return 'number';
+    return 'number'
   }
   if ((object instanceof String) || (typeof object === 'string')) {
-    return 'string';
+    return 'string'
   }
   if ((object instanceof Boolean) || (typeof object === 'boolean')) {
-    return 'boolean';
+    return 'boolean'
   }
-  if ((object instanceof RegExp) || (typeof object === 'regexp')) {
-    return 'regexp';
+  if (object instanceof RegExp) {
+    return 'regexp'
   }
-  if (exports.isArray(object)) {
-    return 'array';
+  if (isArray(object)) {
+    return 'array'
   }
 
-  return 'object';
-};
+  return 'object'
+}
 
 /**
  * Test whether a text contains a url (matches when a string starts
  * with 'http://*' or 'https://*' and has no whitespace characters)
  * @param {String} text
  */
-var isUrlRegex = /^https?:\/\/\S+$/;
-exports.isUrl = function isUrl (text) {
-  return (typeof text == 'string' || text instanceof String) &&
-      isUrlRegex.test(text);
-};
+const isUrlRegex = /^https?:\/\/\S+$/
+
+export function isUrl (text) {
+  return (typeof text === 'string' || text instanceof String) &&
+      isUrlRegex.test(text)
+}
 
 /**
  * Tes whether given object is an Array
  * @param {*} obj
  * @returns {boolean} returns true when obj is an array
  */
-exports.isArray = function (obj) {
-  return Object.prototype.toString.call(obj) === '[object Array]';
-};
+export function isArray (obj) {
+  return Object.prototype.toString.call(obj) === '[object Array]'
+}
 
 /**
  * Retrieve the absolute left value of a DOM element
@@ -272,10 +395,10 @@ exports.isArray = function (obj) {
  * @return {Number} left    The absolute left position of this element
  *                          in the browser page.
  */
-exports.getAbsoluteLeft = function getAbsoluteLeft(elem) {
-  var rect = elem.getBoundingClientRect();
-  return rect.left + window.pageXOffset || document.scrollLeft || 0;
-};
+export function getAbsoluteLeft (elem) {
+  const rect = elem.getBoundingClientRect()
+  return rect.left + window.pageXOffset || document.scrollLeft || 0
+}
 
 /**
  * Retrieve the absolute top value of a DOM element
@@ -283,69 +406,77 @@ exports.getAbsoluteLeft = function getAbsoluteLeft(elem) {
  * @return {Number} top     The absolute top position of this element
  *                          in the browser page.
  */
-exports.getAbsoluteTop = function getAbsoluteTop(elem) {
-  var rect = elem.getBoundingClientRect();
-  return rect.top + window.pageYOffset || document.scrollTop || 0;
-};
+export function getAbsoluteTop (elem) {
+  const rect = elem.getBoundingClientRect()
+  return rect.top + window.pageYOffset || document.scrollTop || 0
+}
 
 /**
  * add a className to the given elements style
  * @param {Element} elem
  * @param {String} className
  */
-exports.addClassName = function addClassName(elem, className) {
-  var classes = elem.className.split(' ');
-  if (classes.indexOf(className) == -1) {
-    classes.push(className); // add the class to the array
-    elem.className = classes.join(' ');
+export function addClassName (elem, className) {
+  const classes = elem.className.split(' ')
+  if (classes.indexOf(className) === -1) {
+    classes.push(className) // add the class to the array
+    elem.className = classes.join(' ')
   }
-};
+}
+
+/**
+ * remove all classes from the given elements style
+ * @param {Element} elem
+ */
+export function removeAllClassNames (elem) {
+  elem.className = ''
+}
 
 /**
  * add a className to the given elements style
  * @param {Element} elem
  * @param {String} className
  */
-exports.removeClassName = function removeClassName(elem, className) {
-  var classes = elem.className.split(' ');
-  var index = classes.indexOf(className);
-  if (index != -1) {
-    classes.splice(index, 1); // remove the class from the array
-    elem.className = classes.join(' ');
+export function removeClassName (elem, className) {
+  const classes = elem.className.split(' ')
+  const index = classes.indexOf(className)
+  if (index !== -1) {
+    classes.splice(index, 1) // remove the class from the array
+    elem.className = classes.join(' ')
   }
-};
+}
 
 /**
  * Strip the formatting from the contents of a div
  * the formatting from the div itself is not stripped, only from its childs.
  * @param {Element} divElement
  */
-exports.stripFormatting = function stripFormatting(divElement) {
-  var childs = divElement.childNodes;
-  for (var i = 0, iMax = childs.length; i < iMax; i++) {
-    var child = childs[i];
+export function stripFormatting (divElement) {
+  const childs = divElement.childNodes
+  for (let i = 0, iMax = childs.length; i < iMax; i++) {
+    const child = childs[i]
 
     // remove the style
     if (child.style) {
       // TODO: test if child.attributes does contain style
-      child.removeAttribute('style');
+      child.removeAttribute('style')
     }
 
     // remove all attributes
-    var attributes = child.attributes;
+    const attributes = child.attributes
     if (attributes) {
-      for (var j = attributes.length - 1; j >= 0; j--) {
-        var attribute = attributes[j];
+      for (let j = attributes.length - 1; j >= 0; j--) {
+        const attribute = attributes[j]
         if (attribute.specified === true) {
-          child.removeAttribute(attribute.name);
+          child.removeAttribute(attribute.name)
         }
       }
     }
 
     // recursively strip childs
-    exports.stripFormatting(child);
+    stripFormatting(child)
   }
-};
+}
 
 /**
  * Set focus to the end of an editable div
@@ -354,67 +485,67 @@ exports.stripFormatting = function stripFormatting(divElement) {
  * http://stackoverflow.com/questions/1125292/how-to-move-cursor-to-end-of-contenteditable-entity
  * @param {Element} contentEditableElement   A content editable div
  */
-exports.setEndOfContentEditable = function setEndOfContentEditable(contentEditableElement) {
-  var range, selection;
-  if(document.createRange) {
-    range = document.createRange();//Create a range (a range is a like the selection but invisible)
-    range.selectNodeContents(contentEditableElement);//Select the entire contents of the element with the range
-    range.collapse(false);//collapse the range to the end point. false means collapse to end rather than the start
-    selection = window.getSelection();//get the selection object (allows you to change selection)
-    selection.removeAllRanges();//remove any selections already made
-    selection.addRange(range);//make the range you have just created the visible selection
+export function setEndOfContentEditable (contentEditableElement) {
+  let range, selection
+  if (document.createRange) {
+    range = document.createRange()// Create a range (a range is a like the selection but invisible)
+    range.selectNodeContents(contentEditableElement)// Select the entire contents of the element with the range
+    range.collapse(false)// collapse the range to the end point. false means collapse to end rather than the start
+    selection = window.getSelection()// get the selection object (allows you to change selection)
+    selection.removeAllRanges()// remove any selections already made
+    selection.addRange(range)// make the range you have just created the visible selection
   }
-};
+}
 
 /**
  * Select all text of a content editable div.
  * http://stackoverflow.com/a/3806004/1262753
  * @param {Element} contentEditableElement   A content editable div
  */
-exports.selectContentEditable = function selectContentEditable(contentEditableElement) {
-  if (!contentEditableElement || contentEditableElement.nodeName != 'DIV') {
-    return;
+export function selectContentEditable (contentEditableElement) {
+  if (!contentEditableElement || contentEditableElement.nodeName !== 'DIV') {
+    return
   }
 
-  var sel, range;
+  let sel, range
   if (window.getSelection && document.createRange) {
-    range = document.createRange();
-    range.selectNodeContents(contentEditableElement);
-    sel = window.getSelection();
-    sel.removeAllRanges();
-    sel.addRange(range);
+    range = document.createRange()
+    range.selectNodeContents(contentEditableElement)
+    sel = window.getSelection()
+    sel.removeAllRanges()
+    sel.addRange(range)
   }
-};
+}
 
 /**
  * Get text selection
  * http://stackoverflow.com/questions/4687808/contenteditable-selected-text-save-and-restore
  * @return {Range | TextRange | null} range
  */
-exports.getSelection = function getSelection() {
+export function getSelection () {
   if (window.getSelection) {
-    var sel = window.getSelection();
+    const sel = window.getSelection()
     if (sel.getRangeAt && sel.rangeCount) {
-      return sel.getRangeAt(0);
+      return sel.getRangeAt(0)
     }
   }
-  return null;
-};
+  return null
+}
 
 /**
  * Set text selection
  * http://stackoverflow.com/questions/4687808/contenteditable-selected-text-save-and-restore
  * @param {Range | TextRange | null} range
  */
-exports.setSelection = function setSelection(range) {
+export function setSelection (range) {
   if (range) {
     if (window.getSelection) {
-      var sel = window.getSelection();
-      sel.removeAllRanges();
-      sel.addRange(range);
+      const sel = window.getSelection()
+      sel.removeAllRanges()
+      sel.addRange(range)
     }
   }
-};
+}
 
 /**
  * Get selected text range
@@ -425,20 +556,20 @@ exports.setSelection = function setSelection(range) {
  *                                                   selected text element
  *                          Returns null if no text selection is found
  */
-exports.getSelectionOffset = function getSelectionOffset() {
-  var range = exports.getSelection();
+export function getSelectionOffset () {
+  const range = getSelection()
 
   if (range && 'startOffset' in range && 'endOffset' in range &&
-      range.startContainer && (range.startContainer == range.endContainer)) {
+      range.startContainer && (range.startContainer === range.endContainer)) {
     return {
       startOffset: range.startOffset,
       endOffset: range.endOffset,
       container: range.startContainer.parentNode
-    };
+    }
   }
 
-  return null;
-};
+  return null
+}
 
 /**
  * Set selected text range in given element
@@ -447,25 +578,25 @@ exports.getSelectionOffset = function getSelectionOffset() {
  *                              {Number} startOffset
  *                              {Number} endOffset
  */
-exports.setSelectionOffset = function setSelectionOffset(params) {
+export function setSelectionOffset (params) {
   if (document.createRange && window.getSelection) {
-    var selection = window.getSelection();
-    if(selection) {
-      var range = document.createRange();
+    const selection = window.getSelection()
+    if (selection) {
+      const range = document.createRange()
 
       if (!params.container.firstChild) {
-        params.container.appendChild(document.createTextNode(''));
+        params.container.appendChild(document.createTextNode(''))
       }
 
       // TODO: do not suppose that the first child of the container is a textnode,
       //       but recursively find the textnodes
-      range.setStart(params.container.firstChild, params.startOffset);
-      range.setEnd(params.container.firstChild, params.endOffset);
+      range.setStart(params.container.firstChild, params.startOffset)
+      range.setEnd(params.container.firstChild, params.endOffset)
 
-      exports.setSelection(range);
+      setSelection(range)
     }
   }
-};
+}
 
 /**
  * Get the inner text of an HTML element (for example a div element)
@@ -473,70 +604,86 @@ exports.setSelectionOffset = function setSelectionOffset(params) {
  * @param {Object} [buffer]
  * @return {String} innerText
  */
-exports.getInnerText = function getInnerText(element, buffer) {
-  var first = (buffer == undefined);
+export function getInnerText (element, buffer) {
+  const first = (buffer === undefined)
   if (first) {
     buffer = {
-      'text': '',
-      'flush': function () {
-        var text = this.text;
-        this.text = '';
-        return text;
+      text: '',
+      flush: function () {
+        const text = this.text
+        this.text = ''
+        return text
       },
-      'set': function (text) {
-        this.text = text;
+      set: function (text) {
+        this.text = text
       }
-    };
+    }
   }
 
   // text node
   if (element.nodeValue) {
-    return buffer.flush() + element.nodeValue;
+    return buffer.flush() + element.nodeValue
   }
 
   // divs or other HTML elements
   if (element.hasChildNodes()) {
-    var childNodes = element.childNodes;
-    var innerText = '';
+    const childNodes = element.childNodes
+    let innerText = ''
 
-    for (var i = 0, iMax = childNodes.length; i < iMax; i++) {
-      var child = childNodes[i];
+    for (let i = 0, iMax = childNodes.length; i < iMax; i++) {
+      const child = childNodes[i]
 
-      if (child.nodeName == 'DIV' || child.nodeName == 'P') {
-        var prevChild = childNodes[i - 1];
-        var prevName = prevChild ? prevChild.nodeName : undefined;
-        if (prevName && prevName != 'DIV' && prevName != 'P' && prevName != 'BR') {
-          innerText += '\n';
-          buffer.flush();
+      if (child.nodeName === 'DIV' || child.nodeName === 'P') {
+        const prevChild = childNodes[i - 1]
+        const prevName = prevChild ? prevChild.nodeName : undefined
+        if (prevName && prevName !== 'DIV' && prevName !== 'P' && prevName !== 'BR') {
+          innerText += '\n'
+          buffer.flush()
         }
-        innerText += exports.getInnerText(child, buffer);
-        buffer.set('\n');
-      }
-      else if (child.nodeName == 'BR') {
-        innerText += buffer.flush();
-        buffer.set('\n');
-      }
-      else {
-        innerText += exports.getInnerText(child, buffer);
+        innerText += getInnerText(child, buffer)
+        buffer.set('\n')
+      } else if (child.nodeName === 'BR') {
+        innerText += buffer.flush()
+        buffer.set('\n')
+      } else {
+        innerText += getInnerText(child, buffer)
       }
     }
 
-    return innerText;
-  }
-  else {
-    if (element.nodeName == 'P' && exports.getInternetExplorerVersion() != -1) {
+    return innerText
+  } else {
+    if (element.nodeName === 'P' && getInternetExplorerVersion() !== -1) {
       // On Internet Explorer, a <p> with hasChildNodes()==false is
       // rendered with a new line. Note that a <p> with
       // hasChildNodes()==true is rendered without a new line
       // Other browsers always ensure there is a <br> inside the <p>,
       // and if not, the <p> does not render a new line
-      return buffer.flush();
+      return buffer.flush()
     }
   }
 
   // br or unknown
-  return '';
-};
+  return ''
+}
+
+/**
+ * Test whether an element has the provided parent node somewhere up the node tree.
+ * @param {Element} elem
+ * @param {Element} parent
+ * @return {boolean}
+ */
+export function hasParentNode (elem, parent) {
+  let e = elem ? elem.parentNode : undefined
+
+  while (e) {
+    if (e === parent) {
+      return true
+    }
+    e = e.parentNode
+  }
+
+  return false
+}
 
 /**
  * Returns the version of Internet Explorer or a -1
@@ -544,38 +691,37 @@ exports.getInnerText = function getInnerText(element, buffer) {
  * Source: http://msdn.microsoft.com/en-us/library/ms537509(v=vs.85).aspx
  * @return {Number} Internet Explorer version, or -1 in case of an other browser
  */
-exports.getInternetExplorerVersion = function getInternetExplorerVersion() {
-  if (_ieVersion == -1) {
-    var rv = -1; // Return value assumes failure.
-    if (navigator.appName == 'Microsoft Internet Explorer')
-    {
-      var ua = navigator.userAgent;
-      var re  = new RegExp("MSIE ([0-9]{1,}[\.0-9]{0,})");
+export function getInternetExplorerVersion () {
+  if (_ieVersion === -1) {
+    let rv = -1 // Return value assumes failure.
+    if (typeof navigator !== 'undefined' && navigator.appName === 'Microsoft Internet Explorer') {
+      const ua = navigator.userAgent
+      const re = new RegExp('MSIE ([0-9]+[.0-9]+)')
       if (re.exec(ua) != null) {
-        rv = parseFloat( RegExp.$1 );
+        rv = parseFloat(RegExp.$1)
       }
     }
 
-    _ieVersion = rv;
+    _ieVersion = rv
   }
 
-  return _ieVersion;
-};
+  return _ieVersion
+}
 
 /**
  * Test whether the current browser is Firefox
  * @returns {boolean} isFirefox
  */
-exports.isFirefox = function isFirefox () {
-  return (navigator.userAgent.indexOf("Firefox") != -1);
-};
+export function isFirefox () {
+  return (typeof navigator !== 'undefined' && navigator.userAgent.indexOf('Firefox') !== -1)
+}
 
 /**
  * cached internet explorer version
  * @type {Number}
  * @private
  */
-var _ieVersion = -1;
+var _ieVersion = -1
 
 /**
  * Add and event listener. Works for all browsers
@@ -586,26 +732,23 @@ var _ieVersion = -1;
  * @param {boolean}     [useCapture] false by default
  * @return {function}   the created event listener
  */
-exports.addEventListener = function addEventListener(element, action, listener, useCapture) {
+export function addEventListener (element, action, listener, useCapture) {
   if (element.addEventListener) {
-    if (useCapture === undefined)
-      useCapture = false;
+    if (useCapture === undefined) { useCapture = false }
 
-    if (action === "mousewheel" && exports.isFirefox()) {
-      action = "DOMMouseScroll";  // For Firefox
+    if (action === 'mousewheel' && isFirefox()) {
+      action = 'DOMMouseScroll' // For Firefox
     }
 
-    element.addEventListener(action, listener, useCapture);
-    return listener;
+    element.addEventListener(action, listener, useCapture)
+    return listener
   } else if (element.attachEvent) {
     // Old IE browsers
-    var f = function () {
-      return listener.call(element, window.event);
-    };
-    element.attachEvent("on" + action, f);
-    return f;
+    const f = () => listener.call(element, window.event)
+    element.attachEvent('on' + action, f)
+    return f
   }
-};
+}
 
 /**
  * Remove an event listener from an element
@@ -614,84 +757,180 @@ exports.addEventListener = function addEventListener(element, action, listener, 
  * @param {function} listener  The listener function
  * @param {boolean}  [useCapture]   false by default
  */
-exports.removeEventListener = function removeEventListener(element, action, listener, useCapture) {
+export function removeEventListener (element, action, listener, useCapture) {
   if (element.removeEventListener) {
-    if (useCapture === undefined)
-      useCapture = false;
+    if (useCapture === undefined) { useCapture = false }
 
-    if (action === "mousewheel" && exports.isFirefox()) {
-      action = "DOMMouseScroll";  // For Firefox
+    if (action === 'mousewheel' && isFirefox()) {
+      action = 'DOMMouseScroll' // For Firefox
     }
 
-    element.removeEventListener(action, listener, useCapture);
+    element.removeEventListener(action, listener, useCapture)
   } else if (element.detachEvent) {
     // Old IE browsers
-    element.detachEvent("on" + action, listener);
+    element.detachEvent('on' + action, listener)
   }
-};
+}
+
+/**
+ * Test if an element is a child of a parent element.
+ * @param {Element} elem
+ * @param {Element} parent
+ * @return {boolean} returns true if elem is a child of the parent
+ */
+export function isChildOf (elem, parent) {
+  let e = elem.parentNode
+  while (e) {
+    if (e === parent) {
+      return true
+    }
+    e = e.parentNode
+  }
+
+  return false
+}
 
 /**
  * Parse a JSON path like '.items[3].name' into an array
  * @param {string} jsonPath
  * @return {Array}
  */
-exports.parsePath = function parsePath(jsonPath) {
-  var prop, remainder;
+export function parsePath (jsonPath) {
+  const path = []
+  let i = 0
 
-  if (jsonPath.length === 0) {
-    return [];
-  }
-
-  // find a match like '.prop'
-  var match = jsonPath.match(/^\.(\w+)/);
-  if (match) {
-    prop = match[1];
-    remainder = jsonPath.substr(prop.length + 1);
-  }
-  else if (jsonPath[0] === '[') {
-    // find a match like
-    var end = jsonPath.indexOf(']');
-    if (end === -1) {
-      throw new SyntaxError('Character ] expected in path');
-    }
-    if (end === 1) {
-      throw new SyntaxError('Index expected after [');
+  function parseProperty () {
+    let prop = ''
+    while (jsonPath[i] !== undefined && /[\w$]/.test(jsonPath[i])) {
+      prop += jsonPath[i]
+      i++
     }
 
-    prop = JSON.parse(jsonPath.substring(1, end));
-    remainder = jsonPath.substr(end + 1);
-  }
-  else {
-    throw new SyntaxError('Failed to parse path');
+    if (prop === '') {
+      throw new Error('Invalid JSON path: property name expected at index ' + i)
+    }
+
+    return prop
   }
 
-  return [prop].concat(parsePath(remainder))
-};
+  function parseIndex (end) {
+    let name = ''
+    while (jsonPath[i] !== undefined && jsonPath[i] !== end) {
+      name += jsonPath[i]
+      i++
+    }
+
+    if (jsonPath[i] !== end) {
+      throw new Error('Invalid JSON path: unexpected end, character ' + end + ' expected')
+    }
+
+    return name
+  }
+
+  while (jsonPath[i] !== undefined) {
+    if (jsonPath[i] === '.') {
+      i++
+      path.push(parseProperty())
+    } else if (jsonPath[i] === '[') {
+      i++
+
+      if (jsonPath[i] === '\'' || jsonPath[i] === '"') {
+        const end = jsonPath[i]
+        i++
+
+        path.push(parseIndex(end))
+
+        if (jsonPath[i] !== end) {
+          throw new Error('Invalid JSON path: closing quote \' expected at index ' + i)
+        }
+        i++
+      } else {
+        let index = parseIndex(']').trim()
+        if (index.length === 0) {
+          throw new Error('Invalid JSON path: array value expected at index ' + i)
+        }
+        // Coerce numeric indices to numbers, but ignore star
+        index = index === '*' ? index : JSON.parse(index)
+        path.push(index)
+      }
+
+      if (jsonPath[i] !== ']') {
+        throw new Error('Invalid JSON path: closing bracket ] expected at index ' + i)
+      }
+      i++
+    } else {
+      throw new Error('Invalid JSON path: unexpected character "' + jsonPath[i] + '" at index ' + i)
+    }
+  }
+
+  return path
+}
+
+/**
+ * Stringify an array with a path in a JSON path like '.items[3].name'
+ * @param {Array.<string | number>} path
+ * @returns {string}
+ */
+export function stringifyPath (path) {
+  return path
+    .map(p => {
+      if (typeof p === 'number') {
+        return ('[' + p + ']')
+      } else if (typeof p === 'string' && p.match(/^[A-Za-z0-9_$]+$/)) {
+        return '.' + p
+      } else {
+        return '["' + p + '"]'
+      }
+    })
+    .join('')
+}
 
 /**
  * Improve the error message of a JSON schema error
  * @param {Object} error
  * @return {Object} The error
  */
-exports.improveSchemaError = function (error) {
+export function improveSchemaError (error) {
   if (error.keyword === 'enum' && Array.isArray(error.schema)) {
-    var enums = error.schema;
+    let enums = error.schema
     if (enums) {
-      enums = enums.map(function (value) {
-        return JSON.stringify(value);
-      });
+      enums = enums.map(value => JSON.stringify(value))
 
       if (enums.length > 5) {
-        var more = ['(' + (enums.length - 5) + ' more...)'];
-        enums = enums.slice(0, 5);
-        enums.push(more);
+        const more = ['(' + (enums.length - 5) + ' more...)']
+        enums = enums.slice(0, 5)
+        enums.push(more)
       }
-      error.message = 'should be equal to one of: ' + enums.join(', ');
+      error.message = 'should be equal to one of: ' + enums.join(', ')
     }
   }
 
-  return error;
-};
+  if (error.keyword === 'additionalProperties') {
+    error.message = 'should NOT have additional property: ' + error.params.additionalProperty
+  }
+
+  return error
+}
+
+/**
+ * Test whether something is a Promise
+ * @param {*} object
+ * @returns {boolean} Returns true when object is a promise, false otherwise
+ */
+export function isPromise (object) {
+  return object && typeof object.then === 'function' && typeof object.catch === 'function'
+}
+
+/**
+ * Test whether a custom validation error has the correct structure
+ * @param {*} validationError The error to be checked.
+ * @returns {boolean} Returns true if the structure is ok, false otherwise
+ */
+export function isValidValidationError (validationError) {
+  return typeof validationError === 'object' &&
+      Array.isArray(validationError.path) &&
+      typeof validationError.message === 'string'
+}
 
 /**
  * Test whether the child rect fits completely inside the parent rect.
@@ -699,13 +938,13 @@ exports.improveSchemaError = function (error) {
  * @param {ClientRect} child
  * @param {number} margin
  */
-exports.insideRect = function (parent, child, margin) {
-  var _margin = margin !== undefined ? margin : 0;
-  return child.left   - _margin >= parent.left
-      && child.right  + _margin <= parent.right
-      && child.top    - _margin >= parent.top
-      && child.bottom + _margin <= parent.bottom;
-};
+export function insideRect (parent, child, margin) {
+  const _margin = margin !== undefined ? margin : 0
+  return child.left - _margin >= parent.left &&
+      child.right + _margin <= parent.right &&
+      child.top - _margin >= parent.top &&
+      child.bottom + _margin <= parent.bottom
+}
 
 /**
  * Returns a function, that, as long as it continues to be invoked, will not
@@ -721,20 +960,20 @@ exports.insideRect = function (parent, child, margin) {
  *                                      of the trailing.
  * @return {function} Return the debounced function
  */
-exports.debounce = function debounce(func, wait, immediate) {
-  var timeout;
-  return function() {
-    var context = this, args = arguments;
-    var later = function() {
-      timeout = null;
-      if (!immediate) func.apply(context, args);
-    };
-    var callNow = immediate && !timeout;
-    clearTimeout(timeout);
-    timeout = setTimeout(later, wait);
-    if (callNow) func.apply(context, args);
-  };
-};
+export function debounce (func, wait, immediate) {
+  let timeout
+  return function () {
+    const context = this; const args = arguments
+    const later = () => {
+      timeout = null
+      if (!immediate) func.apply(context, args)
+    }
+    const callNow = immediate && !timeout
+    clearTimeout(timeout)
+    timeout = setTimeout(later, wait)
+    if (callNow) func.apply(context, args)
+  }
+}
 
 /**
  * Determines the difference between two texts.
@@ -744,22 +983,481 @@ exports.debounce = function debounce(func, wait, immediate) {
  * @return {{start: number, end: number}} Returns the start and end
  *                                        of the changed part in newText.
  */
-exports.textDiff = function textDiff(oldText, newText) {
-  var len = newText.length;
-  var start = 0;
-  var oldEnd = oldText.length;
-  var newEnd = newText.length;
+export function textDiff (oldText, newText) {
+  const len = newText.length
+  let start = 0
+  let oldEnd = oldText.length
+  let newEnd = newText.length
 
-  while (newText.charAt(start) === oldText.charAt(start)
-  && start < len) {
-    start++;
+  while (newText.charAt(start) === oldText.charAt(start) &&
+  start < len) {
+    start++
   }
 
-  while (newText.charAt(newEnd - 1) === oldText.charAt(oldEnd - 1)
-  && newEnd > start && oldEnd > 0) {
-    newEnd--;
-    oldEnd--;
+  while (newText.charAt(newEnd - 1) === oldText.charAt(oldEnd - 1) &&
+  newEnd > start && oldEnd > 0) {
+    newEnd--
+    oldEnd--
   }
 
-  return {start: start, end: newEnd};
-};
+  return { start: start, end: newEnd }
+}
+
+/**
+ * Return an object with the selection range or cursor position (if both have the same value)
+ * Support also old browsers (IE8-)
+ * Source: http://ourcodeworld.com/articles/read/282/how-to-get-the-current-cursor-position-and-selection-within-a-text-input-or-textarea-in-javascript
+ * @param {DOMElement} el A dom element of a textarea or input text.
+ * @return {Object} reference Object with 2 properties (start and end) with the identifier of the location of the cursor and selected text.
+ **/
+export function getInputSelection (el) {
+  let startIndex = 0; let endIndex = 0; let normalizedValue; let range; let textInputRange; let len; let endRange
+
+  if (typeof el.selectionStart === 'number' && typeof el.selectionEnd === 'number') {
+    startIndex = el.selectionStart
+    endIndex = el.selectionEnd
+  } else {
+    range = document.selection.createRange()
+
+    if (range && range.parentElement() === el) {
+      len = el.value.length
+      normalizedValue = el.value.replace(/\r\n/g, '\n')
+
+      // Create a working TextRange that lives only in the input
+      textInputRange = el.createTextRange()
+      textInputRange.moveToBookmark(range.getBookmark())
+
+      // Check if the startIndex and endIndex of the selection are at the very end
+      // of the input, since moveStart/moveEnd doesn't return what we want
+      // in those cases
+      endRange = el.createTextRange()
+      endRange.collapse(false)
+
+      if (textInputRange.compareEndPoints('StartToEnd', endRange) > -1) {
+        startIndex = endIndex = len
+      } else {
+        startIndex = -textInputRange.moveStart('character', -len)
+        startIndex += normalizedValue.slice(0, startIndex).split('\n').length - 1
+
+        if (textInputRange.compareEndPoints('EndToEnd', endRange) > -1) {
+          endIndex = len
+        } else {
+          endIndex = -textInputRange.moveEnd('character', -len)
+          endIndex += normalizedValue.slice(0, endIndex).split('\n').length - 1
+        }
+      }
+    }
+  }
+
+  return {
+    startIndex: startIndex,
+    endIndex: endIndex,
+    start: _positionForIndex(startIndex),
+    end: _positionForIndex(endIndex)
+  }
+
+  /**
+   * Returns textarea row and column position for certain index
+   * @param {Number} index text index
+   * @returns {{row: Number, column: Number}}
+   */
+  function _positionForIndex (index) {
+    const textTillIndex = el.value.substring(0, index)
+    const row = (textTillIndex.match(/\n/g) || []).length + 1
+    const col = textTillIndex.length - textTillIndex.lastIndexOf('\n')
+
+    return {
+      row: row,
+      column: col
+    }
+  }
+}
+
+/**
+ * Returns the index for certaion position in text element
+ * @param {DOMElement} el A dom element of a textarea or input text.
+ * @param {Number} row row value, > 0, if exceeds rows number - last row will be returned
+ * @param {Number} column column value, > 0, if exceeds column length - end of column will be returned
+ * @returns {Number} index of position in text, -1 if not found
+ */
+export function getIndexForPosition (el, row, column) {
+  const text = el.value || ''
+  if (row > 0 && column > 0) {
+    const rows = text.split('\n', row)
+    row = Math.min(rows.length, row)
+    column = Math.min(rows[row - 1].length, column - 1)
+    const columnCount = (row === 1 ? column : column + 1) // count new line on multiple rows
+    return rows.slice(0, row - 1).join('\n').length + columnCount
+  }
+  return -1
+}
+
+/**
+ * Returns location of json paths in certain json string
+ * @param {String} text json string
+ * @param {Array<String>} paths array of json paths
+ * @returns {Array<{path: String, line: Number, row: Number}>}
+ */
+export function getPositionForPath (text, paths) {
+  const result = []
+  let jsmap
+  if (!paths || !paths.length) {
+    return result
+  }
+
+  try {
+    jsmap = jsonMap.parse(text)
+  } catch (err) {
+    return result
+  }
+
+  paths.forEach(path => {
+    const pathArr = parsePath(path)
+    const pointerName = compileJSONPointer(pathArr)
+    const pointer = jsmap.pointers[pointerName]
+    if (pointer) {
+      result.push({
+        path: path,
+        line: pointer.key ? pointer.key.line : (pointer.value ? pointer.value.line : 0),
+        column: pointer.key ? pointer.key.column : (pointer.value ? pointer.value.column : 0)
+      })
+    }
+  })
+
+  return result
+}
+
+/**
+ * Compile a JSON Pointer
+ * WARNING: this is an incomplete implementation
+ * @param {Array.<string | number>} path
+ * @return {string}
+ */
+export function compileJSONPointer (path) {
+  return path
+    .map(p => ('/' + String(p)
+      .replace(/~/g, '~0')
+      .replace(/\//g, '~1')
+    ))
+    .join('')
+}
+
+/**
+ * Get the applied color given a color name or code
+ * Source: https://stackoverflow.com/questions/6386090/validating-css-color-names/33184805
+ * @param {string} color
+ * @returns {string | null} returns the color if the input is a valid
+ *                   color, and returns null otherwise. Example output:
+ *                   'rgba(255,0,0,0.7)' or 'rgb(255,0,0)'
+ */
+export function getColorCSS (color) {
+  const ele = document.createElement('div')
+  ele.style.color = color
+  return ele.style.color.split(/\s+/).join('').toLowerCase() || null
+}
+
+/**
+ * Test if a string contains a valid color name or code.
+ * @param {string} color
+ * @returns {boolean} returns true if a valid color, false otherwise
+ */
+export function isValidColor (color) {
+  return !!getColorCSS(color)
+}
+
+/**
+ * Make a tooltip for a field based on the field's schema.
+ * @param {object} schema JSON schema
+ * @param {string} [locale] Locale code (for example, zh-CN)
+ * @returns {string} Field tooltip, may be empty string if all relevant schema properties are missing
+ */
+export function makeFieldTooltip (schema, locale) {
+  if (!schema) {
+    return ''
+  }
+
+  let tooltip = ''
+  if (schema.title) {
+    tooltip += schema.title
+  }
+
+  if (schema.description) {
+    if (tooltip.length > 0) {
+      tooltip += '\n'
+    }
+    tooltip += schema.description
+  }
+
+  if (schema.default) {
+    if (tooltip.length > 0) {
+      tooltip += '\n\n'
+    }
+    tooltip += translate('default', undefined, locale) + '\n'
+    tooltip += JSON.stringify(schema.default, null, 2)
+  }
+
+  if (Array.isArray(schema.examples) && schema.examples.length > 0) {
+    if (tooltip.length > 0) {
+      tooltip += '\n\n'
+    }
+    tooltip += translate('examples', undefined, locale) + '\n'
+    schema.examples.forEach((example, index) => {
+      tooltip += JSON.stringify(example, null, 2)
+      if (index !== schema.examples.length - 1) {
+        tooltip += '\n'
+      }
+    })
+  }
+
+  return tooltip
+}
+
+/**
+ * Get a nested property from an object.
+ * Returns undefined when the property does not exist.
+ * @param {Object} object
+ * @param {string[]} path
+ * @return {*}
+ */
+export function get (object, path) {
+  let value = object
+
+  for (let i = 0; i < path.length && value !== undefined && value !== null; i++) {
+    value = value[path[i]]
+  }
+
+  return value
+}
+
+/**
+ * Find a unique name. Suffix the name with ' (copy)', '(copy 2)', etc
+ * until a unique name is found
+ * @param {string} name
+ * @param {Array} existingPropNames    Array with existing prop names
+ */
+export function findUniqueName (name, existingPropNames) {
+  const strippedName = name.replace(/ \(copy( \d+)?\)$/, '')
+  let validName = strippedName
+  let i = 1
+
+  while (existingPropNames.indexOf(validName) !== -1) {
+    const copy = 'copy' + (i > 1 ? (' ' + i) : '')
+    validName = strippedName + ' (' + copy + ')'
+    i++
+  }
+
+  return validName
+}
+
+/**
+ * Get the child paths of an array
+ * @param {JSON} json
+ * @param {boolean} [includeObjects=false] If true, object and array paths are returned as well
+ * @return {string[]}
+ */
+export function getChildPaths (json, includeObjects) {
+  const pathsMap = {}
+
+  function getObjectChildPaths (json, pathsMap, rootPath, includeObjects) {
+    const isValue = !Array.isArray(json) && !isObject(json)
+
+    if (isValue || includeObjects) {
+      pathsMap[rootPath || ''] = true
+    }
+
+    if (isObject(json)) {
+      Object.keys(json).forEach(field => {
+        getObjectChildPaths(json[field], pathsMap, rootPath + '.' + field, includeObjects)
+      })
+    }
+  }
+
+  if (Array.isArray(json)) {
+    const max = Math.min(json.length, MAX_ITEMS_FIELDS_COLLECTION)
+    for (let i = 0; i < max; i++) {
+      const item = json[i]
+      getObjectChildPaths(item, pathsMap, '', includeObjects)
+    }
+  } else {
+    pathsMap[''] = true
+  }
+
+  return Object.keys(pathsMap).sort()
+}
+
+/**
+ * Sort object keys using natural sort
+ * @param {Array} array
+ * @param {String} [path] JSON pointer
+ * @param {'asc' | 'desc'} [direction]
+ */
+export function sort (array, path, direction) {
+  const parsedPath = path && path !== '.' ? parsePath(path) : []
+  const sign = direction === 'desc' ? -1 : 1
+
+  const sortedArray = array.slice()
+  sortedArray.sort((a, b) => {
+    const aValue = get(a, parsedPath)
+    const bValue = get(b, parsedPath)
+
+    return sign * (aValue > bValue ? 1 : aValue < bValue ? -1 : 0)
+  })
+
+  return sortedArray
+}
+
+/**
+ * Sort object keys using natural sort
+ * @param {Object} object
+ * @param {'asc' | 'desc'} [direction]
+ */
+export function sortObjectKeys (object, direction) {
+  const sign = (direction === 'desc') ? -1 : 1
+  const sortedFields = Object.keys(object).sort((a, b) => sign * naturalSort(a, b))
+
+  const sortedObject = {}
+  sortedFields.forEach(field => {
+    sortedObject[field] = object[field]
+  })
+
+  return sortedObject
+}
+
+/**
+ * Cast contents of a string to the correct type.
+ * This can be a string, a number, a boolean, etc
+ * @param {String} str
+ * @return {*} castedStr
+ * @private
+ */
+export function parseString (str) {
+  if (str === '') {
+    return ''
+  }
+
+  const lower = str.toLowerCase()
+  if (lower === 'null') {
+    return null
+  }
+  if (lower === 'true') {
+    return true
+  }
+  if (lower === 'false') {
+    return false
+  }
+
+  const num = Number(str) // will nicely fail with '123ab'
+  const numFloat = parseFloat(str) // will nicely fail with '  '
+  if (!isNaN(num) && !isNaN(numFloat)) {
+    return num
+  }
+
+  return str
+}
+
+/**
+ * Test whether some field contains a timestamp in milliseconds after the year 2000.
+ * @param {string} field
+ * @param {number} value
+ * @return {boolean}
+ */
+export function isTimestamp (field, value) {
+  return typeof value === 'number' &&
+    value > YEAR_2000 &&
+    isFinite(value) &&
+    Math.floor(value) === value &&
+    !isNaN(new Date(value).valueOf())
+}
+
+/**
+ * Return a human readable document size
+ * For example formatSize(7570718) outputs '7.2 MiB'
+ * @param {number} size
+ * @return {string} Returns a human readable size
+ */
+export function formatSize (size) {
+  if (size < 900) {
+    return size.toFixed() + ' B'
+  }
+
+  const KiB = size / 1024
+  if (KiB < 900) {
+    return KiB.toFixed(1) + ' KiB'
+  }
+
+  const MiB = KiB / 1024
+  if (MiB < 900) {
+    return MiB.toFixed(1) + ' MiB'
+  }
+
+  const GiB = MiB / 1024
+  if (GiB < 900) {
+    return GiB.toFixed(1) + ' GiB'
+  }
+
+  const TiB = GiB / 1024
+  return TiB.toFixed(1) + ' TiB'
+}
+
+/**
+ * Limit text to a maximum number of characters
+ * @param {string} text
+ * @param {number} maxCharacterCount
+ * @return {string} Returns the limited text,
+ *                  ending with '...' if the max was exceeded
+ */
+export function limitCharacters (text, maxCharacterCount) {
+  if (text.length <= maxCharacterCount) {
+    return text
+  }
+
+  return text.slice(0, maxCharacterCount) + '...'
+}
+
+/**
+ * Test whether a value is an Object
+ * @param {*} value
+ * @return {boolean}
+ */
+export function isObject (value) {
+  return typeof value === 'object' && value !== null && !Array.isArray(value)
+}
+
+/**
+ * Helper function to test whether an array contains an item
+ * @param {Array} array
+ * @param {*} item
+ * @return {boolean} Returns true if `item` is in `array`, returns false otherwise.
+ */
+export function contains (array, item) {
+  return array.indexOf(item) !== -1
+}
+
+/**
+ * Checkes if validation has changed from the previous execution
+ * @param {Array} currErr current validation errors
+ * @param {Array} prevErr previous validation errors
+ */
+export function isValidationErrorChanged (currErr, prevErr) {
+  if (!prevErr && !currErr) { return false }
+  if ((prevErr && !currErr) || (!prevErr && currErr)) { return true }
+  if (prevErr.length !== currErr.length) { return true }
+
+  for (let i = 0; i < currErr.length; ++i) {
+    let pErr
+    if (currErr[i].type === 'error') {
+      pErr = prevErr.find(p => p.line === currErr[i].line)
+    } else {
+      pErr = prevErr.find(p => p.dataPath === currErr[i].dataPath && p.schemaPath === currErr[i].schemaPath)
+    }
+    if (!pErr) {
+      return true
+    }
+  }
+
+  return false
+}
+
+function hasOwnProperty (object, key) {
+  return Object.prototype.hasOwnProperty.call(object, key)
+}
