@@ -1,38 +1,54 @@
-var fs = require('fs');
-var gulp = require('gulp');
-var gutil = require('gulp-util');
-var concatCss = require('gulp-concat-css');
-var minifyCSS = require('gulp-minify-css');
-var shell = require('gulp-shell');
-var mkdirp = require('mkdirp');
-var webpack = require('webpack');
-var uglify = require('uglify-js');
+const fs = require('fs')
+const path = require('path')
+const gulp = require('gulp')
+const log = require('fancy-log')
+const format = require('date-format')
+const concatCss = require('gulp-concat-css')
+const minifyCSS = require('gulp-clean-css')
+const sass = require('gulp-sass')
+const mkdirp = require('mkdirp')
+const webpack = require('webpack')
+const uglify = require('uglify-js')
+const btoa = require('btoa')
 
-var NAME    = 'jsoneditor';
-var NAME_MINIMALIST = 'jsoneditor-minimalist';
-var ENTRY   = './src/js/JSONEditor.js';
-var HEADER  = './src/js/header.js';
-var IMAGE   = './src/css/img/jsoneditor-icons.svg';
-var DOCS    = './src/docs/*';
-var DIST    = './dist';
+const NAME = 'jsoneditor'
+const NAME_MINIMALIST = 'jsoneditor-minimalist'
+const ENTRY = './src/js/JSONEditor.js'
+const HEADER = './src/js/header.js'
+const IMAGE = './src/scss/img/jsoneditor-icons.svg'
+const DOCS = './src/docs/*'
+const DIST = path.join(__dirname, 'dist')
 
 // generate banner with today's date and correct version
-function createBanner() {
-  var today = gutil.date(new Date(), 'yyyy-mm-dd'); // today, formatted as yyyy-mm-dd
-  var version = require('./package.json').version;  // math.js version
+function createBanner () {
+  const today = format.asString('yyyy-MM-dd', new Date()) // today, formatted as yyyy-MM-dd
+  const version = require('./package.json').version // math.js version
 
   return String(fs.readFileSync(HEADER))
-      .replace('@@date', today)
-      .replace('@@version', version);
+    .replace('@@date', today)
+    .replace('@@version', version)
 }
 
-var bannerPlugin = new webpack.BannerPlugin(createBanner(), {
+const bannerPlugin = new webpack.BannerPlugin({
+  banner: createBanner(),
   entryOnly: true,
   raw: true
-});
+})
+
+const webpackConfigModule = {
+  rules: [
+    {
+      test: /\.m?js$/,
+      exclude: /node_modules/,
+      use: {
+        loader: 'babel-loader'
+      }
+    }
+  ]
+}
 
 // create a single instance of the compiler to allow caching
-var compiler = webpack({
+const compiler = webpack({
   entry: ENTRY,
   output: {
     library: 'JSONEditor',
@@ -40,17 +56,21 @@ var compiler = webpack({
     path: DIST,
     filename: NAME + '.js'
   },
-  plugins: [ bannerPlugin ],
-  module: {
-    loaders: [
-      { test: /\.json$/, loader: "json" }
-    ]
+  plugins: [bannerPlugin],
+  optimization: {
+    // We no not want to minimize our code.
+    minimize: false
+  },
+  module: webpackConfigModule,
+  resolve: {
+    extensions: ['.js'],
+    mainFields: ['main'] // pick ES5 version of vanilla-picker
   },
   cache: true
-});
+})
 
 // create a single instance of the compiler to allow caching
-var compilerMinimalist = webpack({
+const compilerMinimalist = webpack({
   entry: ENTRY,
   output: {
     library: 'JSONEditor',
@@ -58,130 +78,172 @@ var compilerMinimalist = webpack({
     path: DIST,
     filename: NAME_MINIMALIST + '.js'
   },
+  module: webpackConfigModule,
   plugins: [
     bannerPlugin,
-    new webpack.IgnorePlugin(new RegExp('^brace$')),
-    new webpack.IgnorePlugin(new RegExp('^ajv'))
+    new webpack.IgnorePlugin(new RegExp('^ace-builds')),
+    new webpack.IgnorePlugin(new RegExp('worker-json-data-url')),
+    new webpack.IgnorePlugin(new RegExp('^ajv')),
+    new webpack.IgnorePlugin(new RegExp('^vanilla-picker'))
   ],
+  optimization: {
+    // We no not want to minimize our code.
+    minimize: false
+  },
   cache: true
-});
+})
 
-function minify(name) {
-  var result = uglify.minify([DIST + '/' + name + '.js'], {
-    outSourceMap: name + '.map',
+function minify (name) {
+  const code = String(fs.readFileSync(DIST + '/' + name + '.js'))
+  const result = uglify.minify(code, {
+    sourceMap: {
+      url: name + '.map'
+    },
     output: {
-      comments: /@license/
+      comments: /@license/,
+      max_line_len: 64000 // extra large because we have embedded code for workers
     }
-  });
+  })
 
-  var fileMin = DIST + '/' + name + '.min.js';
-  var fileMap = DIST + '/' + name + '.map';
+  if (result.error) {
+    throw result.error
+  }
 
-  fs.writeFileSync(fileMin, result.code);
-  fs.writeFileSync(fileMap, result.map);
+  const fileMin = DIST + '/' + name + '.min.js'
+  const fileMap = DIST + '/' + name + '.map'
 
-  gutil.log('Minified ' + fileMin);
-  gutil.log('Mapped ' + fileMap);
+  fs.writeFileSync(fileMin, result.code)
+  fs.writeFileSync(fileMap, result.map)
+
+  log('Minified ' + fileMin)
+  log('Mapped ' + fileMap)
 }
 
-// make dist and dist/img folders
-gulp.task('mkdir', function () {
-  mkdirp.sync(DIST);
-  mkdirp.sync(DIST + '/img');
-});
+// make dist folder structure
+gulp.task('mkdir', function (done) {
+  mkdirp.sync(DIST)
+  mkdirp.sync(DIST + '/img')
+
+  done()
+})
+
+// Create an embedded version of the json worker code: a data url
+gulp.task('embed-json-worker', function (done) {
+  const workerBundleFile = './node_modules/ace-builds/src-noconflict/worker-json.js'
+  const workerEmbeddedFile = './src/js/generated/worker-json-data-url.js'
+  const workerScript = String(fs.readFileSync(workerBundleFile))
+
+  const workerDataUrl = 'data:application/javascript;base64,' + btoa(workerScript)
+
+  fs.writeFileSync(workerEmbeddedFile, 'module.exports = \'' + workerDataUrl + '\'\n')
+
+  done()
+})
 
 // bundle javascript
-gulp.task('bundle', ['mkdir'], function (done) {
+gulp.task('bundle', function (done) {
   // update the banner contents (has a date in it which should stay up to date)
-  bannerPlugin.banner = createBanner();
+  bannerPlugin.banner = createBanner()
 
   compiler.run(function (err, stats) {
     if (err) {
-      gutil.log(err);
+      log(err)
     }
 
-    gutil.log('bundled ' + NAME + '.js');
+    log('bundled ' + NAME + '.js')
 
-    done();
-  });
-});
+    done()
+  })
+})
 
 // bundle minimalist version of javascript
-gulp.task('bundle-minimalist', ['mkdir'], function (done) {
+gulp.task('bundle-minimalist', function (done) {
   // update the banner contents (has a date in it which should stay up to date)
-  bannerPlugin.banner = createBanner();
+  bannerPlugin.banner = createBanner()
 
   compilerMinimalist.run(function (err, stats) {
     if (err) {
-      gutil.log(err);
+      log(err)
     }
 
-    gutil.log('bundled ' + NAME_MINIMALIST + '.js');
+    log('bundled ' + NAME_MINIMALIST + '.js')
 
-    done();
-  });
-});
+    done()
+  })
+})
 
 // bundle css
-gulp.task('bundle-css', ['mkdir'], function () {
-  gulp.src([
-    'src/css/jsoneditor.css',
-    'src/css/contextmenu.css',
-    'src/css/menu.css',
-    'src/css/searchbox.css'
-  ])
-      .pipe(concatCss(NAME + '.css'))
-      .pipe(gulp.dest(DIST))
-      .pipe(concatCss(NAME + '.min.css'))
-      .pipe(minifyCSS())
-      .pipe(gulp.dest(DIST));
-
-  gutil.log('bundled ' + DIST + '/' + NAME + '.css');
-  gutil.log('bundled ' + DIST + '/' + NAME + '.min.css');
-});
+gulp.task('bundle-css', function (done) {
+  gulp
+    .src([
+      'src/scss/reset.scss',
+      'src/scss/jsoneditor.scss',
+      'src/scss/contextmenu.scss',
+      'src/scss/menu.scss',
+      'src/scss/searchbox.scss',
+      'src/scss/autocomplete.scss',
+      'src/scss/treepath.scss',
+      'src/scss/statusbar.scss',
+      'src/scss/navigationbar.scss',
+      'src/js/assets/selectr/selectr.scss'
+    ])
+    .pipe(
+      sass({
+        // importer: tildeImporter
+      })
+    )
+    .pipe(concatCss(NAME + '.css'))
+    .pipe(gulp.dest(DIST))
+    .pipe(concatCss(NAME + '.min.css'))
+    .pipe(minifyCSS())
+    .pipe(gulp.dest(DIST))
+  done()
+})
 
 // create a folder img and copy the icons
-gulp.task('copy-img', ['mkdir'], function () {
-  gulp.src(IMAGE)
-      .pipe(gulp.dest(DIST +'/img'));
-  gutil.log('Copied images');
-});
+gulp.task('copy-img', function (done) {
+  gulp.src(IMAGE).pipe(gulp.dest(DIST + '/img'))
+  log('Copied images')
+
+  done()
+})
 
 // create a folder img and copy the icons
-gulp.task('copy-docs', ['mkdir'], function () {
-  gulp.src(DOCS)
-      .pipe(gulp.dest(DIST));
-  gutil.log('Copied doc');
-});
+gulp.task('copy-docs', function (done) {
+  gulp.src(DOCS).pipe(gulp.dest(DIST))
+  log('Copied doc')
 
-gulp.task('minify', ['bundle'], function () {
+  done()
+})
+
+gulp.task('minify', function (done) {
   minify(NAME)
-});
 
-gulp.task('minify-minimalist', ['bundle-minimalist'], function () {
+  done()
+})
+
+gulp.task('minify-minimalist', function (done) {
   minify(NAME_MINIMALIST)
-});
 
-// TODO: zip file using archiver
-var pkg = 'jsoneditor-' + require('./package.json').version + '.zip';
-gulp.task('zip', shell.task([
-      'zip ' + pkg + ' ' + 'README.md NOTICE LICENSE HISTORY.md index.html src dist docs examples -r '
-]));
+  done()
+})
 
 // The watch task (to automatically rebuild when the source code changes)
 // Does only generate jsoneditor.js and jsoneditor.css, and copy the image
 // Does NOT minify the code and does NOT generate the minimalist version
-gulp.task('watch', ['bundle', 'bundle-css', 'copy-img'], function () {
-  gulp.watch(['src/**/*'], ['bundle', 'bundle-css', 'copy-img']);
-});
+gulp.task('watch', gulp.series('bundle', 'bundle-css', 'copy-img', function () {
+  gulp.watch(['src/**/*'], gulp.series('bundle', 'bundle-css', 'copy-img'))
+}))
 
 // The default task (called when you run `gulp`)
-gulp.task('default', [
-  'bundle',
-  'bundle-minimalist',
-  'bundle-css',
-  'copy-img',
-  'copy-docs',
-  'minify',
-  'minify-minimalist'
-]);
+gulp.task('default', gulp.series(
+  'mkdir',
+  'embed-json-worker',
+  gulp.parallel(
+    'copy-img',
+    'copy-docs',
+    'bundle-css',
+    gulp.series('bundle', 'minify'),
+    gulp.series('bundle-minimalist', 'minify-minimalist')
+  )
+))
